+++ conf/imsconfig.go
@@ -177,7 +177,6 @@ type ConfigCore struct {
 	Port                 int32
 	AccessTokenLifetime  time.Duration
 	RefreshTokenLifetime time.Duration
-	Admins               []string
 	MasterKey            string `redact:"true"`
 	// #nosec G117 // Exported secret struct field
 	JWTSecret  string `redact:"true"`